refactor!: type Accessor.Min/Max as List<double> (#716)
Wider element type preserves precision for the double component type
introduced by the forthcoming glTF 2.1 specification and avoids a
lossy round-trip for values exceeding float precision.

BREAKING CHANGE: Accessor.Min and Accessor.Max are now List<double>
instead of List<float>. Update assignments and reads accordingly.

+++ Packages/com.unity.cloud.gltfast.tests/Tests/Runtime/Scripts/Import/GltfBufferMock.cs
@@ -21,8 +21,8 @@ namespace GLTFast.Tests
                 Sparse = index == sparseAccessorIndex
                     ? new AccessorSparse()
                     : null,
-                Min = new List<float> { -1, -1, -1 },
-                Max = new List<float> { 1, 1, 1 }
+                Min = new List<double> { -1, -1, -1 },
+                Max = new List<double> { 1, 1, 1 }
             };
         }
