+++ Packages/com.unity.cloud.gltfast/Runtime/Scripts/Export/GltfWriter.cs
@@ -1155,8 +1155,8 @@ namespace GLTFast.Export
                         var bounds = uMesh.bounds;
                         var max = bounds.max;
                         var min = bounds.min;
-                        accessor.Min = new List<float> { -max.x, min.y, min.z };
-                        accessor.Max = new List<float> { -min.x, max.y, max.z };
+                        accessor.Min = new List<double> { -max.x, min.y, min.z };
+                        accessor.Max = new List<double> { -min.x, max.y, max.z };
                         attributes.Position = accessorId;
                         break;
                     case VertexAttribute.Normal:
@@ -1649,13 +1649,13 @@ namespace GLTFast.Export
                         var bounds = submeshDesc.bounds;
                         var center = bounds.center;
                         var extents = bounds.extents;
-                        accessor.Min = new List<float>
+                        accessor.Min = new List<double>
                         {
                             -center.x-extents.x,
                             center.y-extents.y,
                             center.z-extents.z
                         };
-                        accessor.Max = new List<float>
+                        accessor.Max = new List<double>
                         {
                             -center.x+extents.x,
                             center.y+extents.y,
